requirements.txt
openai
langchain_openai
neo4j-driver
kneed==0.8.5
langchain==0.2.3
langchain_community==0.2.4
langchain_core==0.2.5
numpy==1.26.4
opencv_python==4.9.0.80
overrides==7.7.0
Pillow==10.3.0
Requests==2.32.3
scikit_learn==1.4.2
sentence_transformers==2.7.0
streamlit==1.33.0
tqdm==4.66.2
